% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getResource.R
\docType{data}
\name{FormioInterface}
\alias{FormioInterface}
\title{Extract all results of a given resource or form}
\format{An object of class \code{R6ClassGenerator} of length 24.}
\usage{
FormioInterface
}
\arguments{
\item{base_url}{Base formio URL}

\item{resource_path}{Name of the path of the form ie: 'user'.}

\item{token}{The formio token to allow access}

\item{filter}{The list of filters}

\item{filter}{The list of items to include in the select}

\item{chunk_size}{How many elements to extract, usefull for large Data sets}

\item{filters}{the list of filters}

\item{select}{the list of select}

\item{limit}{the limit for the query}

\item{size}{Number of elements on each call}
}
\description{
This function allows you to express your love of cats.
}
\examples{
FormioInterface()
Sets the proper Filter URL for the GET CALL

Parses the list of filters to a single string
filterToString()
Sets the proper Select URL for the GET CALL

Parses the list of select elements to a single string
selectToString()
Extract all results of a given resource or form

This function allows you to express your love of cats.
get()
Applies the filters

filter()
Applies the select

select()
Applies the limit to the Query

limit()
Defines how many elements to call each time

limit()
}
\keyword{filter}
\keyword{formio}
