% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getResource.R
\docType{data}
\name{Formior}
\alias{Formior}
\title{Extract all results of a given resource or form}
\format{An object of class \code{R6ClassGenerator} of length 24.}
\usage{
Formior
}
\arguments{
\item{base_url}{Base formio URL}

\item{resource_path}{Name of the path of the form ie: 'user'.}

\item{token}{The formio token to allow access}

\item{filter}{The list of filters}

\item{filter}{The list of items to include in the select}

\item{filter}{The list of items to include in the select}

\item{filters}{the list of filters}

\item{select}{the list of select}

\item{limit}{the limit for the query}

\item{resources}{the populate resources for the query}

\item{size}{Number of elements on each call}
}
\description{
This function allows you to express your love of cats.
}
\examples{
Formior()
Sets the proper Filter URL for the GET CALL

Parses the list of filters to a single string
filterToString()
Sets the proper Select URL for the GET CALL

Parses the list of select elements to a single string
selectToString()
Sets the proper Populate URL for the GET CALL

Parses the list of select elements to a single string
selectToString()
Gets the first Query String

Returns the first query string given all the filters, select and limit
getChunkSize()
Gets the first chunk size

Returns the first chunk size depending on the limit and chunks defined
getChunkSize()
Extract all results of a given resource or form

This function pulls the records
get()
Gets the current selection and flattens the dataframe to a single dimension

This function flattens and returns a dataframe.
flatten()
Applies the filters

filter()
Applies the select

select()
Applies the limit to the Query

limit()
Applies the populate to the Query

limit()
Defines how many elements to call each time

limit()
}
\keyword{filter}
\keyword{formio}
